chore: move all the files into knights folder
