+++ src/handlers/knights/create/schema.validator.ts
