+++ src/handlers/knights/delete/app.ts
@@ -1,8 +1,8 @@
 import { type APIGatewayEvent, type APIGatewayProxyResult } from 'aws-lambda'
-import { BadRequest, MalformedObject, Ok } from '../../infra/helpers/http.helper'
-import { isNullOrUndefined } from '../../infra/helpers/verification.helper'
+import { BadRequest, MalformedObject, Ok } from '../../../infra/helpers/http.helper'
+import { isNullOrUndefined } from '../../../infra/helpers/verification.helper'
 import { routeParamValidator } from './schema.validator'
-import KnightServiceFactory from '../../main/factories/knight-service.factory'
+import KnightServiceFactory from '../../../main/factories/knight-service.factory'
 
 export const handle = async (event: APIGatewayEvent): Promise<APIGatewayProxyResult> => {
   const params = event.pathParameters as { id: string }
