+++ src/handlers/knights/delete/schema.validator.ts
