+++ src/handlers/knights/get-by-id/schema.validator.ts
