+++ src/handlers/knights/get/schema.validator.ts
