+++ src/handlers/knights/update/schema.validator.ts
